+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [1.20.4-2.2.1.0] - 2024-05-11
 ### Added
-- better compatibility with mods that do not support ItemHandlers (NeoForge)
+- better compatibility with mods that do not support ItemHandlers (Forge & NeoForge)
 
 ## [1.20.4-2.2.0.1] - 2024-04-24
 ### Fixed